Complile Sass to CSS using liveCompiler
Switched the initial workflow and setup for this project from gulp, to
a less complicated setup using the liveCompilers vsCode extension to
compile Sass files to CSS, minify the generated css and create source
maps. The main reason for the switch was because i was experiencing
issues with the generated source maps for the minified css and
javascript files pointing to the wrong line numbers.