Add a range inputfor password length
Refactored the HTML code to include a range input to handle user
password length selection. The slider's min value is 6 and max value is
20 and an initial start value of 6.

Also refactored the code in app.js. Within the intiApp function, the
code was simplified make sure at least one character group is selected
before the form can be submitted.

A new function was aslo created to give visual feedback to the user on
the current value of the slider.