Style the recent password widget
completed the styling for the widget that showa a list of the five most
recent password generated. I also made some minor adjustment to the font
-size of the title element of the widget. worked on the styling for the
copy button on each password and the clear history button. made some
minor adjustments to the title of the project and the margin left values
of the custom _checkboxes so they can align properly with the text on
smaller devices. The with of the main copy button (yellow) was adjusted
to be longer on smaller screens and 50% shorter on larger screens.